libnodectld: stop transfer subprocess groups on rollback
Track transfer subprocesses as their own process groups and terminate
the whole group during rollback so interrupted sends and receives do
not leave helper processes behind.

+++ libnodectld/lib/nodectld/utils/subprocess.rb
@@ -7,20 +7,36 @@ module NodeCtld
     # on *this* node until that subprocess finishes.
     # It has no effect on transactions on other nodes.
     def blocking_fork(&)
-      child = Process.fork(&)
+      child = Process.fork do
+        Process.setpgrp
+        yield
+      end
+
       Daemon.register_subprocess(@command.chain_id, child)
     end
 
     def killall_subprocesses
-      @daemon.chain_blockers do |blockers|
+      daemon = @daemon || Daemon.instance
+      return unless daemon
+
+      daemon.chain_blockers do |blockers|
         next unless blockers
 
         log('Killing all subprocesses')
 
         blockers.each_value do |pids|
           pids.each do |pid|
-            log("Sending SIGTERM to subprocess #{pid}")
-            Process.kill('TERM', pid)
+            log("Sending SIGTERM to subprocess group #{pid}")
+
+            begin
+              Process.kill('TERM', -pid)
+            rescue Errno::ESRCH
+              begin
+                Process.kill('TERM', pid)
+              rescue Errno::ESRCH
+                nil
+              end
+            end
           end
         end
       end
